# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tawjihi Practice Exams API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specializations:
    get:
      operationId: getSpecializations
      tags: [specializations]
      summary: List all specializations
      responses:
        "200":
          description: List of specializations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Specialization"
    post:
      operationId: createSpecialization
      tags: [specializations]
      summary: Create a specialization (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecializationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialization"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /specializations/{id}:
    put:
      operationId: updateSpecialization
      tags: [specializations]
      summary: Update a specialization (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecializationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialization"
    delete:
      operationId: deleteSpecialization
      tags: [specializations]
      summary: Delete a specialization (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: List subjects by specialization
      parameters:
        - name: specializationId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /units:
    get:
      operationId: getUnits
      tags: [units]
      summary: List units by subject
      parameters:
        - name: subjectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    put:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, supervisor, student, guest]
        provider:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - createdAt
...
    Specialization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateSpecializationBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specializationId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - specializationId
        - createdAt
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        specializationId:
          type: integer
      required:
        - name
        - specializationId
    Unit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subjectId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - subjectId
        - createdAt
    CreateUnitBody:
      type: object
      properties:
        name:
          type: string
        subjectId:
          type: integer
      required:
        - name
        - subjectId
...
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, supervisor, student, guest]
        name:
          type: string
      required:
        - role
...