# lib/api-zod/src/generated/types/unit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tawjihi Practice Exams API
 * OpenAPI spec version: 0.1.0
 */

export interface Unit {
  id: number;
  name: string;
  subjectId: number;
  createdAt: Date;
}